pyproject.toml
[project]
name = "compliance-rag"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "chromadb>=1.5.1",
    "faiss-cpu>=1.13.2",
    "fastembed>=0.7.4",
    "ipykernel>=7.2.0",
    "langchain>=1.2.10",
    "langchain-cohere>=0.5.0",
    "langchain-community>=0.4.1",
    "langchain-core>=1.2.16",
    "langchain-groq>=1.1.2",
    "langchain-huggingface>=1.2.0",
    "langchain-openai>=1.1.10",
    "langchain-qdrant>=1.1.0",
    "psycopg2>=2.9.11",
    "pymupdf>=1.27.1",
    "pypdf>=6.7.4",
    "python-dotenv>=1.2.1",
    "tiktoken>=0.12.0",
]
